+++ src/Scene.ts
@@ -361,6 +361,13 @@ export class Scene implements StorageLoaderHandler {
 	operation: Trigger<OperationEvent>;
 
 	/**
+	 * ゲーム開発者向けのコンテナ。
+	 *
+	 * この値はゲームエンジンのロジックからは使用されず、ゲーム開発者は任意の目的に使用してよい。
+	 */
+	vars: any;
+
+	/**
 	 * @private
 	 */
 	_storageLoader: StorageLoader | undefined;
@@ -430,6 +437,11 @@ export class Scene implements StorageLoaderHandler {
 	_assetHolders: AssetHolder<SceneRequestAssetHandler>[];
 
 	/**
+	 * @private
+	 */
+	_waitingPrepare: ((done: () => void) => void) | undefined;
+
+	/**
 	 * 各種パラメータを指定して `Scene` のインスタンスを生成する。
 	 * @param param 初期化に用いるパラメータのオブジェクト
 	 */
@@ -464,6 +476,7 @@ export class Scene implements StorageLoaderHandler {
 		this._ready = this._onReady;
 		this.assets = {};
 		this.asset = new AssetAccessor(game._assetManager);
+		this.vars = {};
 
 		this._loaded = false;
 		this._prefetchRequested = false;
@@ -558,6 +571,7 @@ export class Scene implements StorageLoaderHandler {
 		this.onAssetLoadFailure.destroy();
 		this.onAssetLoadComplete.destroy();
 		this.assets = {};
+		this.vars = {};
 
 		// アセットを参照しているEより先に解放しないよう最後に解放する
 		for (let i = 0; i < this._assetHolders.length; ++i) this._assetHolders[i].destroy();
@@ -566,6 +580,7 @@ export class Scene implements StorageLoaderHandler {
 		this._storageLoader = undefined;
 
 		this.game = undefined!;
+		this._waitingPrepare = undefined;
 
 		this.state = "destroyed";
 		this.onStateChange.fire(this.state);
@@ -858,7 +873,11 @@ export class Scene implements StorageLoaderHandler {
 	 * @private
 	 */
 	_needsLoading(): boolean {
-		return this._sceneAssetHolder.waitingAssetsCount > 0 || (!!this._storageLoader && !this._storageLoader._loaded);
+		return (
+			this._sceneAssetHolder.waitingAssetsCount > 0 ||
+			(!!this._storageLoader && !this._storageLoader._loaded) ||
+			!!this._waitingPrepare
+		);
 	}
 
 	/**